feat: move edit modal to its own component, minor changes
